Charges Glories
Currently only supports Wildy Sword 4 and an ornate pool in house, near portal. Does not withdraw house tabs, must have as many as you plan to use for duration of the script. Will run until out of glories or out of house tabs.  
Wider support coming, was just excited to have a first functional script  and wanted to share :D
1. Have Wildy Sword 4 and a glory equipped. 
2. Have house tabs anywhere at the bottom of the inventory.
3. Have uncharged glories filling remaining inventory. 
4. Have uncharged glories visible at the top of the bank .